move artifact suite into separate package
